+++ src/sensors.cu
@@ -7,11 +7,11 @@ MTL::MTL()
     this->julian_day = 0;
     this->number_sensor = 0;
     this->sun_elevation = 0;
-    this->rad_add = (float *)malloc(7 * sizeof(float));
-    this->rad_mult = (float *)malloc(7 * sizeof(float));
-    this->ref_add = (float *)malloc(7 * sizeof(float));
-    this->ref_mult = (float *)malloc(7 * sizeof(float));
-    this->ref_w_coeff = (float *)malloc(7 * sizeof(float));
+    this->rad_add = (double *)malloc(7 * sizeof(double));
+    this->rad_mult = (double *)malloc(7 * sizeof(double));
+    this->ref_add = (double *)malloc(7 * sizeof(double));
+    this->ref_mult = (double *)malloc(7 * sizeof(double));
+    this->ref_w_coeff = (double *)malloc(7 * sizeof(double));
 };
 
 MTL::MTL(string metadata_path)
@@ -60,26 +60,26 @@ MTL::MTL(string metadata_path)
     this->distance_earth_sun = atof(mtl["EARTH_SUN_DISTANCE"].c_str());
     this->image_hour = (hours + minutes / 60.0) * 100;
 
-    this->rad_add = (float *)malloc(7 * sizeof(float));
-    this->rad_mult = (float *)malloc(7 * sizeof(float));
-    this->ref_add = (float *)malloc(7 * sizeof(float));
-    this->ref_mult = (float *)malloc(7 * sizeof(float));
-    this->ref_w_coeff = (float *)malloc(7 * sizeof(float));
+    this->rad_add = (double *)malloc(7 * sizeof(double));
+    this->rad_mult = (double *)malloc(7 * sizeof(double));
+    this->ref_add = (double *)malloc(7 * sizeof(double));
+    this->ref_mult = (double *)malloc(7 * sizeof(double));
+    this->ref_w_coeff = (double *)malloc(7 * sizeof(double));
 
-    HANDLE_ERROR(cudaMalloc((void **)&this->rad_add_d, (7 * sizeof(float))));
-    HANDLE_ERROR(cudaMalloc((void **)&this->rad_mult_d, (7 * sizeof(float))));
-    HANDLE_ERROR(cudaMalloc((void **)&this->ref_add_d, (7 * sizeof(float))));
-    HANDLE_ERROR(cudaMalloc((void **)&this->ref_mult_d, (7 * sizeof(float))));
-    HANDLE_ERROR(cudaMalloc((void **)&this->ref_w_coeff_d, (7 * sizeof(float))));
+    HANDLE_ERROR(cudaMalloc((void **)&this->rad_add_d, (7 * sizeof(double))));
+    HANDLE_ERROR(cudaMalloc((void **)&this->rad_mult_d, (7 * sizeof(double))));
+    HANDLE_ERROR(cudaMalloc((void **)&this->ref_add_d, (7 * sizeof(double))));
+    HANDLE_ERROR(cudaMalloc((void **)&this->ref_mult_d, (7 * sizeof(double))));
+    HANDLE_ERROR(cudaMalloc((void **)&this->ref_w_coeff_d, (7 * sizeof(double))));
 
     if (this->number_sensor == 8) {
-        this->ref_w_coeff[PARAM_BAND_BLUE_INDEX] = 0.257048331f;
-        this->ref_w_coeff[PARAM_BAND_GREEN_INDEX] = 0.251150748f;
-        this->ref_w_coeff[PARAM_BAND_RED_INDEX] = 0.220943613f;
-        this->ref_w_coeff[PARAM_BAND_NIR_INDEX] = 0.143411968f;
-        this->ref_w_coeff[PARAM_BAND_SWIR1_INDEX] = 0.116657077f;
-        this->ref_w_coeff[PARAM_BAND_TERMAL_INDEX] = 0.000000000f;
-        this->ref_w_coeff[PARAM_BAND_SWIR2_INDEX] = 0.010788262f;
+        this->ref_w_coeff[PARAM_BAND_BLUE_INDEX] = 0.257048331;
+        this->ref_w_coeff[PARAM_BAND_GREEN_INDEX] = 0.251150748;
+        this->ref_w_coeff[PARAM_BAND_RED_INDEX] = 0.220943613;
+        this->ref_w_coeff[PARAM_BAND_NIR_INDEX] = 0.143411968;
+        this->ref_w_coeff[PARAM_BAND_SWIR1_INDEX] = 0.116657077;
+        this->ref_w_coeff[PARAM_BAND_TERMAL_INDEX] = 0.000000000;
+        this->ref_w_coeff[PARAM_BAND_SWIR2_INDEX] = 0.010788262;
 
         this->rad_mult[PARAM_BAND_BLUE_INDEX] = atof(mtl["RADIANCE_MULT_BAND_2"].c_str());
         this->rad_mult[PARAM_BAND_GREEN_INDEX] = atof(mtl["RADIANCE_MULT_BAND_3"].c_str());
@@ -113,13 +113,13 @@ MTL::MTL(string metadata_path)
         this->ref_add[PARAM_BAND_TERMAL_INDEX] = atof(mtl["REFLECTANCE_ADD_BAND_10"].c_str());
         this->ref_add[PARAM_BAND_SWIR2_INDEX] = atof(mtl["REFLECTANCE_ADD_BAND_7"].c_str());
     } else {
-        this->ref_w_coeff[PARAM_BAND_BLUE_INDEX] = 0.298220602f;
-        this->ref_w_coeff[PARAM_BAND_GREEN_INDEX] = 0.270097933f;
-        this->ref_w_coeff[PARAM_BAND_RED_INDEX] = 0.230996896f;
-        this->ref_w_coeff[PARAM_BAND_NIR_INDEX] = 0.155050651f;
-        this->ref_w_coeff[PARAM_BAND_SWIR1_INDEX] = 0.033085493f;
-        this->ref_w_coeff[PARAM_BAND_TERMAL_INDEX] = 0.000000000f;
-        this->ref_w_coeff[PARAM_BAND_SWIR2_INDEX] = 0.012548425f;
+        this->ref_w_coeff[PARAM_BAND_BLUE_INDEX] = 0.298220602;
+        this->ref_w_coeff[PARAM_BAND_GREEN_INDEX] = 0.270097933;
+        this->ref_w_coeff[PARAM_BAND_RED_INDEX] = 0.230996896;
+        this->ref_w_coeff[PARAM_BAND_NIR_INDEX] = 0.155050651;
+        this->ref_w_coeff[PARAM_BAND_SWIR1_INDEX] = 0.033085493;
+        this->ref_w_coeff[PARAM_BAND_TERMAL_INDEX] = 0.000000000;
+        this->ref_w_coeff[PARAM_BAND_SWIR2_INDEX] = 0.012548425;
 
         this->rad_mult[PARAM_BAND_BLUE_INDEX] = atof(mtl["RADIANCE_MULT_BAND_1"].c_str());
         this->rad_mult[PARAM_BAND_GREEN_INDEX] = atof(mtl["RADIANCE_MULT_BAND_2"].c_str());
@@ -154,11 +154,11 @@ MTL::MTL(string metadata_path)
         this->ref_add[PARAM_BAND_SWIR2_INDEX] = atof(mtl["REFLECTANCE_ADD_BAND_7"].c_str());
     }
 
-    HANDLE_ERROR(cudaMemcpy(this->rad_add_d, this->rad_add, 7 * sizeof(float), cudaMemcpyHostToDevice));
-    HANDLE_ERROR(cudaMemcpy(this->rad_mult_d, this->rad_mult, 7 * sizeof(float), cudaMemcpyHostToDevice));
-    HANDLE_ERROR(cudaMemcpy(this->ref_add_d, this->ref_add, 7 * sizeof(float), cudaMemcpyHostToDevice));
-    HANDLE_ERROR(cudaMemcpy(this->ref_mult_d, this->ref_mult, 7 * sizeof(float), cudaMemcpyHostToDevice));
-    HANDLE_ERROR(cudaMemcpy(this->ref_w_coeff_d, this->ref_w_coeff, 7 * sizeof(float), cudaMemcpyHostToDevice));
+    HANDLE_ERROR(cudaMemcpy(this->rad_add_d, this->rad_add, 7 * sizeof(double), cudaMemcpyHostToDevice));
+    HANDLE_ERROR(cudaMemcpy(this->rad_mult_d, this->rad_mult, 7 * sizeof(double), cudaMemcpyHostToDevice));
+    HANDLE_ERROR(cudaMemcpy(this->ref_add_d, this->ref_add, 7 * sizeof(double), cudaMemcpyHostToDevice));
+    HANDLE_ERROR(cudaMemcpy(this->ref_mult_d, this->ref_mult, 7 * sizeof(double), cudaMemcpyHostToDevice));
+    HANDLE_ERROR(cudaMemcpy(this->ref_w_coeff_d, this->ref_w_coeff, 7 * sizeof(double), cudaMemcpyHostToDevice));
 };
 
 Station::Station()
@@ -233,7 +233,7 @@ Landsat::Landsat(string bands_paths[])
     TIFFGetField(landsat_bands[1], TIFFTAG_SAMPLEFORMAT, &sample_bands);
 };
 
-void saveTiff(string path, float *data, int height, int width)
+void saveTiff(string path, double *data, int height, int width)
 {
     TIFF *tif = TIFFOpen(path.c_str(), "w");
     TIFFSetField(tif, TIFFTAG_IMAGEWIDTH, width);
@@ -254,7 +254,7 @@ void saveTiff(string path, float *data, int height, int width)
     TIFFClose(tif);
 }
 
-void printLinearPointer(float *pointer, int height, int width)
+void printLinearPointer(double *pointer, int height, int width)
 {
     for (int i = 0; i < height; i++) {
         for (int j = 0; j < width; j++) {
